fixed static reference namespace initialization to be done only once, all top-level QoreProgram objects created contain copies of that namespace, similar to how namespace initialization works with modules fixed that all parse options are respected in the initial program, including "no-system-classes", etc
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@3272 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ include/qore/intern/QoreClassList.h
@@ -52,7 +52,7 @@ public:
    DLLLOCAL ~QoreClassList();
    DLLLOCAL int add(QoreClass *ot);
    DLLLOCAL QoreClass *find(const char *name);
-   DLLLOCAL QoreClassList *copy(int po);
+   DLLLOCAL QoreClassList *copy(int64 po);
    DLLLOCAL void resolveCopy();
    DLLLOCAL void parseInit();
    DLLLOCAL void parseRollback();
